I've cleaned up the repository and moved the project to the root. I deleted all files and directories from the root of the repository, except for the `lagos-oracle-ultra` project folder. I then moved the contents of that folder to the root directory to make it the main project.
