+++ components/ChatInput.tsx
@@ -1,6 +1,6 @@
 
-import React, { useRef, useEffect } from 'react';
-import { Send, Mic, Square } from 'lucide-react';
+import React, { useRef } from 'react';
+import { Send, Mic, MicOff } from 'lucide-react';
 import { ThemeColors } from '../types';
 
 interface ChatInputProps {
@@ -9,9 +9,8 @@ interface ChatInputProps {
   isListening: boolean;
   recognitionAvailable: boolean;
   theme: ThemeColors;
-  setInput: (value: string) => void;
+  setInput: (input: string) => void;
   handleSend: () => void;
-  handleStopGenerating: () => void;
   startListening: () => void;
 }
 
@@ -23,17 +22,9 @@ const ChatInput: React.FC<ChatInputProps> = ({
   theme,
   setInput,
   handleSend,
-  handleStopGenerating,
   startListening
 }) => {
-  const inputRef = useRef<HTMLTextAreaElement>(null);
-
-  useEffect(() => {
-    if (inputRef.current) {
-      inputRef.current.style.height = 'auto';
-      inputRef.current.style.height = `${inputRef.current.scrollHeight}px`;
-    }
-  }, [input]);
+  const inputRef = useRef<HTMLInputElement>(null);
 
   const onSend = () => {
     if (input.trim() && !isTyping) {
@@ -48,49 +39,47 @@ const ChatInput: React.FC<ChatInputProps> = ({
   };
 
   return (
-    <div className={`relative p-4 ${theme.bg} border-t ${theme.muted}`}>
-      <div className={`max-w-4xl mx-auto flex items-end gap-2 ${theme.input} rounded-3xl pr-2 border ${theme.muted} shadow-lg`}>
-        <button
-          onClick={onMicClick}
-          disabled={!recognitionAvailable || isListening || isTyping}
-          className={`p-3 rounded-full transition-colors duration-200 ${isListening ? `${theme.secondaryAccent} text-${theme.text} animate-pulse` : `text-${theme.text} hover:${theme.primaryAccent}`} ${!recognitionAvailable || isTyping ? 'opacity-50 cursor-not-allowed' : ''}`}
-          aria-label={isListening ? "Stop listening" : "Start voice input"}
-        >
-          <Mic className="w-5 h-5" />
-        </button>
-        <textarea
-          ref={inputRef}
-          value={input}
-          onChange={(e) => setInput(e.target.value)}
-          onKeyDown={(e) => {
-            if (e.key === 'Enter' && !e.shiftKey) {
-              e.preventDefault();
-              onSend();
-            }
-          }}
-          placeholder="Message Lagos Oracle..."
-          className={`flex-1 px-4 py-2.5 bg-transparent text-${theme.text} resize-none max-h-40 focus:outline-none placeholder-${theme.muted}`}
-          rows={1}
-          disabled={isTyping}
-        />
-        {!isTyping ? (
+    <div className={`${theme.card} border-t p-3 md:p-4 sticky bottom-0 shadow- ऊपर-md`}> {/* Custom shadow for top */}
+      <div className="max-w-6xl mx-auto">
+        <div className="flex items-center space-x-2">
+          <button
+            onClick={onMicClick}
+            disabled={!recognitionAvailable || isListening}
+            title={recognitionAvailable ? (isListening ? "Listening..." : "Voice Input") : "Voice input not available"}
+            className={`p-3 rounded-xl transition-all duration-200
+              ${isListening ? 'bg-red-500 text-white animate-pulse' :
+              `${theme.muted} ${recognitionAvailable ? 'hover:text-cyan-400 hover:bg-cyan-500/10' : 'opacity-50 cursor-not-allowed'}`}
+            `}
+          >
+            {isListening ? <MicOff className="w-5 h-5" /> : <Mic className="w-5 h-5" />}
+          </button>
+
+          <div className="flex-1 relative">
+            <input
+              ref={inputRef}
+              type="text"
+              value={input}
+              onChange={(e) => setInput(e.target.value)}
+              onKeyPress={(e) => e.key === 'Enter' && onSend()}
+              placeholder="Ask Lagos Oracle anything... (e.g., 'Paint Victoria Island sunset')"
+              className={`w-full px-4 py-3 rounded-xl border ${theme.input}
+                         focus:ring-2 focus:ring-cyan-400/50 focus:border-cyan-400/80
+                         transition-colors duration-200 placeholder-gray-500 dark:placeholder-gray-400/70`}
+              disabled={isTyping}
+            />
+          </div>
+
           <button
             onClick={onSend}
-            className={`p-3 rounded-full transition-colors duration-200 ${input.trim() ? `${theme.primaryAccent} hover:${theme.secondaryAccent}` : `${theme.muted} cursor-not-allowed`} text-${theme.text}`}
-            disabled={!input.trim()}
-            aria-label="Send message"
+            disabled={!input.trim() || isTyping}
+            title="Send Message"
+            className="bg-gradient-to-r from-cyan-400 to-purple-500 text-white p-3 rounded-xl
+                       hover:shadow-lg hover:from-cyan-500 hover:to-purple-600
+                       transition-all duration-200 disabled:opacity-60 disabled:cursor-not-allowed disabled:shadow-none"
           >
             <Send className="w-5 h-5" />
           </button>
-        ) : (
-          <button
-            onClick={handleStopGenerating}
-            className={`p-3 rounded-full transition-colors duration-200 ${theme.secondaryAccent} hover:${theme.secondaryAccent} text-${theme.text}`}
-            aria-label="Stop generating response"
-          >
-            <Square className="w-5 h-5" />
-          </button>
-        )}
+        </div>
       </div>
     </div>
   );
