--- lagos-oracle-ultra/index.html
@@ -1,46 +1,0 @@
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Lagos Oracle Ultra</title>
-  <script src="https://cdn.tailwindcss.com"></script>
-  <style>
-    /* For custom scrollbar styling, optional */
-    ::-webkit-scrollbar {
-      width: 8px;
-    }
-    ::-webkit-scrollbar-track {
-      background: transparent;
-    }
-    ::-webkit-scrollbar-thumb {
-      background: #4f46e5; /* Indigo-600, adjust as needed */
-      border-radius: 4px;
-    }
-    ::-webkit-scrollbar-thumb:hover {
-      background: #4338ca; /* Indigo-700 */
-    }
-    body {
-      overscroll-behavior-y: contain; /* Prevents pull-to-refresh on mobile if content is scrollable */
-    }
-  </style>
-<script type="importmap">
-{
-  "imports": {
-    "react": "https://esm.sh/react@^19.1.0",
-    "react/": "https://esm.sh/react@^19.1.0/",
-    "react-dom/": "https://esm.sh/react-dom@^19.1.0/",
-    "@google/genai": "https://esm.sh/@google/genai@^1.6.0",
-    "lucide-react": "https://esm.sh/lucide-react@^0.522.0",
-    "marked": "https://esm.sh/marked@^15.0.12"
-  }
-}
-</script>
-<link rel="stylesheet" href="/index.css">
-</head>
-<body>
-  <div id="root"></div>
-  <script type="module" src="/index.tsx"></script>
-</body>
-</html>
