--- lagos-oracle-ultra/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "lagos-oracle-ultra",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "@google/genai": "^1.6.0",
-    "lucide-react": "^0.522.0",
-    "marked": "^15.0.12"
-  },
-  "devDependencies": {
-    "@types/node": "^22.14.0",
-    "typescript": "~5.7.2",
-    "vite": "^6.2.0"
-  }
-}
