+++ package.json
@@ -5,42 +5,19 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "build": "vite build",
     "preview": "vite preview"
   },
   "dependencies": {
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "@google/genai": "^1.6.0",
-    "@types/react-window": "^1.8.8",
-    "dist": "^0.1.2",
-    "esm": "^3.2.25",
     "lucide-react": "^0.522.0",
-    "prism": "^4.1.2",
-    "prism-react-renderer": "^2.4.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-markdown": "^9.0.1",
-    "react-router-dom": "^7.6.3",
-    "react-window": "^1.8.11",
-    "remark-gfm": "^4.0.0",
-    "styles": "^0.2.1"
+    "marked": "^15.0.12"
   },
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.13",
     "@types/node": "^22.14.0",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
-    "@types/react-syntax-highlighter": "^15.5.13",
-    "@typescript-eslint/eslint-plugin": "^7.9.0",
-    "@typescript-eslint/parser": "^7.9.0",
-    "@vitejs/plugin-react": "^4.3.0",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.7",
-    "postcss": "^8.4.38",
-    "tailwindcss": "^3.4.3",
     "typescript": "~5.7.2",
-    "vite": "^7.0.0"
+    "vite": "^6.2.0"
   }
 }
